Separate the singer check into an individual check.
