+++ osu.Game.Rulesets.Karaoke/Edit/Checks/CheckLyricLanguage.cs
@@ -9,32 +9,28 @@ using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Checks
 {
-    public class CheckInvalidPropertyLyrics : ICheck
+    public class CheckLyricLanguage : ICheck
     {
-        public CheckMetadata Metadata => new(CheckCategory.HitObjects, "Lyrics with invalid property.");
+        public CheckMetadata Metadata => new(CheckCategory.HitObjects, "Lyrics with invalid language.");
 
         public IEnumerable<IssueTemplate> PossibleTemplates => new IssueTemplate[]
         {
-            new IssueTemplateNoSinger(this),
+            new IssueTemplateNotFillLanguage(this),
         };
 
         public IEnumerable<Issue> Run(BeatmapVerifierContext context)
         {
             foreach (var lyric in context.Beatmap.HitObjects.OfType<Lyric>())
             {
-                // todo : check lyric layout.
-
-                if (!lyric.Singers.Any())
-                    yield return new IssueTemplateNoSinger(this).Create(lyric);
-
-                // todo : check is singer in singer list.
+                if (lyric.Language == null)
+                    yield return new IssueTemplateNotFillLanguage(this).Create(lyric);
             }
         }
 
-        public class IssueTemplateNoSinger : IssueTemplate
+        public class IssueTemplateNotFillLanguage : IssueTemplate
         {
-            public IssueTemplateNoSinger(ICheck check)
-                : base(check, IssueType.Problem, "Lyric must have at least one singer.")
+            public IssueTemplateNotFillLanguage(ICheck check)
+                : base(check, IssueType.Problem, "Lyric must have assign language.")
             {
             }
